% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\docType{class}
\name{divsum-class}
\alias{divsum-class}
\alias{divsum}
\title{The divsum Class}
\description{
The divsum object contents the divsum output from ReaptMasker script calcDivergenceFromAlign.pl
}
\section{Slots}{

\describe{
\item{\code{divergence}}{Weighted average Kimura divergence for each repeat family}

\item{\code{coverage}}{Coverage for each repeat class and divergence (Kimura)}

\item{\code{species}}{species full name}

\item{\code{genome_size}}{genome size}
}}

\concept{objects}
