% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{set_divsum}
\alias{set_divsum}
\title{Creat divsum object}
\usage{
set_divsum(file, species, genome_size)
}
\arguments{
\item{file}{divsum file}

\item{species}{species full name}

\item{genome_size}{genome size}
}
\value{
Returns a meta_data object
}
\description{
creat divsum object
}
\concept{objects}
