% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{show,divsum-method}
\alias{show,divsum-method}
\title{show method for S4 class divsum}
\usage{
\S4method{show}{divsum}(object)
}
\arguments{
\item{object}{An object of class divsum}
}
\description{
show method for S4 class divsum
}
