% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visulization.R
\name{p3}
\alias{p3}
\title{p3
p3}
\usage{
p3(
  object,
  sublabel = FALSE,
  keep_unknown = TRUE,
  width = 0.5,
  palette = "Accent"
)
}
\arguments{
\item{object}{divsum object}

\item{sublabel}{sublabel TRUE/FALSE}

\item{keep_unknown}{keep unknown TRUE/FALSE}

\item{width}{bar width}

\item{palette}{color set: Accent, Dark2, Paired, Pastel1, Pastel2, Set1, Set2, Set3}
}
\description{
p3
p3
}
